# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ShieldOS Privacy Backend API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/vpn/provision:
    post:
      operationId: provisionDevice
      tags: [vpn]
      summary: Provision a new VPN device and generate WireGuard config
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisionDeviceBody"
      responses:
        "201":
          description: Device provisioned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceProvisionResponse"
  /v1/vpn/configs:
    get:
      operationId: listDevices
      tags: [vpn]
      summary: List all user devices
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Device list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceListResponse"
  /v1/vpn/revoke/{deviceId}:
    delete:
      operationId: revokeDevice
      tags: [vpn]
      summary: Revoke a device's VPN access
      security:
        - bearerAuth: []
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /v1/vpn/status:
    get:
      operationId: getVpnStatus
      tags: [vpn]
      summary: Get VPN server status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: VPN status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VpnStatus"

  # BLOCKLIST
...
  /v1/threats/feed:
    get:
      operationId: getThreatFeed
      tags: [threats]
      summary: Get latest verified threats (paginated)
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: verified
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Threat feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreatFeedResponse"
  /v1/threats/report:
    post:
      operationId: reportThreat
      tags: [threats]
      summary: Submit a community threat report
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportThreatBody"
      responses:
        "201":
          description: Threat reported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreatReport"
  /v1/threats/vote/{threatId}:
    post:
      operationId: voteThreat
      tags: [threats]
      summary: Upvote or downvote a threat report
      security:
        - bearerAuth: []
      parameters:
        - name: threatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteBody"
      responses:
        "200":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreatReport"
  /v1/threats/stats:
    get:
      operationId: getThreatStats
      tags: [threats]
      summary: Get threat intelligence statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Threat stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreatStats"

  # DASHBOARD
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
...
    ProvisionDeviceBody:
      type: object
      properties:
        deviceName:
          type: string
          minLength: 1
          maxLength: 64
      required: [deviceName]
    DeviceProvisionResponse:
      type: object
      properties:
        device:
          $ref: "#/components/schemas/Device"
        configContent:
          type: string
          description: WireGuard .conf file content
      required: [device, configContent]
    Device:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        publicKey:
          type: string
        createdAt:
          type: string
          format: date-time
        isActive:
          type: boolean
        blockedCount:
          type: integer
      required: [id, name, publicKey, createdAt, isActive, blockedCount]
    DeviceListResponse:
      type: object
      properties:
        devices:
          type: array
          items:
            $ref: "#/components/schemas/Device"
        total:
          type: integer
      required: [devices, total]
    VpnStatus:
      type: object
      properties:
        activeDevices:
          type: integer
        totalDevices:
          type: integer
        serverUptime:
          type: string
      required: [activeDevices, totalDevices, serverUptime]

    # BLOCKLIST
...
    CategoryCount:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
      required: [category, count]
...
    ThreatReport:
      type: object
      properties:
        id:
          type: integer
        domain:
          type: string
        category:
          type: string
        description:
          type: string
        votes:
          type: integer
        verified:
          type: boolean
        reportedAt:
          type: string
          format: date-time
        userVote:
          type: integer
          description: "1 for upvote, -1 for downvote, 0 for no vote"
      required: [id, domain, category, description, votes, verified, reportedAt, userVote]
    ThreatFeedResponse:
      type: object
      properties:
        threats:
          type: array
          items:
            $ref: "#/components/schemas/ThreatReport"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [threats, total, page, limit]
    ReportThreatBody:
      type: object
      properties:
        domain:
          type: string
        category:
          type: string
          enum: [malware, phishing, tracking, adware, spyware, ransomware, other]
        description:
          type: string
          maxLength: 500
      required: [domain, category, description]
    VoteBody:
      type: object
      properties:
        vote:
          type: integer
          enum: [1, -1]
      required: [vote]
    ThreatStats:
      type: object
      properties:
        totalReports:
          type: integer
        verifiedThreats:
          type: integer
        pendingReview:
          type: integer
        topCategories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
      required: [totalReports, verifiedThreats, pendingReview, topCategories]

    # DASHBOARD
...